Refactor feature card component; move to home page section and update styles for analysis page layout
