+++ src/app/analysis/page.module.css
@@ -1,0 +1,6 @@
+.page {
+  min-height: 80vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
